Pass CVImageBuffers into GPUI instead of IOSurfaces
