+++ crates/media/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "io_surface"
+name = "media"
 version = "0.1.0"
 edition = "2021"
 
 [lib]
-path = "src/io_surface.rs"
+path = "src/media.rs"
 doctest = false
 
 [dependencies]
